:sparkles: Feature(command): add in-file image one-click upload and replace function & add zh-cn sup

+++ esbuild.mjs
@@ -53,7 +53,9 @@ const resultHandler = async (result) => {
 const outdir = './dist'
 
 // clean old built files
-fse.rmdirSync(outdir, { recursive: true })
+if (fse.existsSync(outdir)) {
+  fse.rmdirSync(outdir, { recursive: true })
+}
 
 /** @type {import('esbuild').BuildOptions} */
 const commonOptions = {
